Added Menu to the left side
This script generates a menu out of all existing pages and adds links
to the menu. The links can be creates though creating one file per link
to /config/menu/pages with a url inside the file. The title of the file
will be the link text on the page.
This script is run by the script that generates the index.html